Review Comments fixes

+++ src/main/java/colruyt/rearulmgtdmnejb/bo/RefSourceTypeBo.java
@@ -4,25 +4,18 @@ import java.io.Serializable;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-
-
-/**
- * @version 1.0
- * @created 28-nov-2018 9:14:37
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class RefSourceTypeBo implements Serializable {
 
-	
 	private static final long serialVersionUID = 1L;
-	private long sourceTypeId;
+	private int sourceTypeId;
 	private String sourceName;
 
 	public long getSourceTypeId() {
 		return sourceTypeId;
 	}
 
-	public void setSourceTypeId(long sourceTypeId) {
+	public void setSourceTypeId(int sourceTypeId) {
 		this.sourceTypeId = sourceTypeId;
 	}
